# account-service/src/main/java/com/banka1/account_service/dto/request/ApproveDto.java
package com.banka1.account_service.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@NoArgsConstructor
@AllArgsConstructor
@Getter
@Setter
public class ApproveDto {
    @NotBlank(message = "Unesi key")
    private String key;
}